requirements.txt
flask
flask-cors
faiss-cpu
sentence-transformers
PyMuPDF
torch
transformers
numpy
